requirements.txt
configobj
seaborn
matplotlib
tensorflow-gpu
numpy
pandas
numba
mayavi
opencv-python
tqdm
scipy
plotly